requirements.txt
requests==2.32.3
fastapi===0.115.8
uvicorn==0.33.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
redis==5.2.1